詳細画面のロジックをカスタムフック化
